Allows user to play a 3-D piano by tracking hand movements (uses Leap Motion and Tkinter)
This application contains 3 main features: a virtual piano that users can play either with their hands (in the air) or with a mouse, a page to practice intervals, and a melody playback page. Users can also record and play back their recordings, which are saved even after they exit the application.